corrected placement of stub class
